Add files via upload
Based on the optimal option you find in (a):
(i) Suppose there is an annual cost of $10,000 million to operate any ships (this fixed cost will not vary
with the number of shipping lines kept open). What is the optimal transportation plan?
(ii) The management is considering whether it would be beneficial to expand the supply capacity. The expansion requires an investment of $20,000 per ton of textile. Which of the planets is worth expanding?
(iii) How much the total transportation costs will increase if there is an additional ton of demand at any of the market centres?