Add files via upload
Suppose the management does not need to operate all the five manufacturing plants. Five candidate locations for planets with various operating costs (in $10,000) are given in Table 5.

Candidate Planets A B C D E
Operating Cost 500 400 550 800 600

(i) Given the flexibility of choosing planets, how many and which of the planets should be used?
(ii) How should the management distribute textiles from the planets to the customer locations?
(iii) What are the total cost, transportation cost and operating cost?
(iv) If all the planets need to be used, what would be the total cost, transportation cost and operating cost?